// src/app/models/data.model.ts
export interface IData {
  position?: number;
  ID: string;
  Date: Date;
  Location: ILatLng;
  Param1: number;
  Param2: number;
  Param3: number;
  Param4: number;
  Param5: number;
  Param6: number;
  Param7: string;
  Param8: number;
  Param9: number;
}
export interface  ILatLng {
  Latitude: number;
  Longitude: number;
}